requirements.txt
Django==1.11.10
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
psycopg2==2.7.3.2
PyJWT==1.5.3
pytz==2017.3